+++ src/main/java/org/example/hwshop/domain/Disk.java
@@ -1,0 +1,19 @@
+package org.example.hwshop.domain;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import com.fasterxml.jackson.annotation.JsonProperty;
+import lombok.*;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+@EqualsAndHashCode(callSuper = true)
+@JsonIgnoreProperties(ignoreUnknown = true)
+public class Disk extends Hardware {
+
+    @JsonProperty("Size")
+    private String size;
+
+    @JsonProperty("Port")
+    private String port;
+}